Move chimera-embedding source code under polynomialembeddor.py
