+++ dwave/embedding/polynomialembeddor.py
